sharpen tests up
