+++ sql/analyzer/optimization_rules_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/dolthub/go-mysql-server/memory"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
-	"github.com/dolthub/go-mysql-server/sql/expression/function/aggregation"
 	"github.com/dolthub/go-mysql-server/sql/plan"
 )
 
@@ -446,173 +445,4 @@ func TestRemoveUnnecessaryConverts(t *testing.T) {
 			require.Equal(tt.expected, resultExpr)
 		})
 	}
-}
-
-func TestConvertCrossJoin(t *testing.T) {
-	tableA := memory.NewTable("a", sql.Schema{
-		{Name: "x", Type: sql.Int64, Source: "a"},
-		{Name: "y", Type: sql.Int64, Source: "a"},
-		{Name: "z", Type: sql.Int64, Source: "a"},
-	})
-	tableB := memory.NewTable("b", sql.Schema{
-		{Name: "x", Type: sql.Int64, Source: "b"},
-		{Name: "y", Type: sql.Int64, Source: "b"},
-		{Name: "z", Type: sql.Int64, Source: "b"},
-	})
-
-	fieldAx := expression.NewGetFieldWithTable(0, sql.Int64, "a", "x", false)
-	fieldBy := expression.NewGetFieldWithTable(0, sql.Int64, "b", "y", false)
-	litOne := expression.NewLiteral(1, sql.Int64)
-
-	matching := []sql.Expression{
-		expression.NewEquals(fieldAx, fieldBy),
-		expression.NewNullSafeEquals(fieldAx, fieldBy),
-		expression.NewGreaterThan(fieldAx, fieldBy),
-		expression.NewGreaterThanOrEqual(fieldAx, fieldBy),
-		expression.NewNullSafeGreaterThan(fieldAx, fieldBy),
-		expression.NewNullSafeGreaterThanOrEqual(fieldAx, fieldBy),
-		expression.NewLessThan(fieldAx, fieldBy),
-		expression.NewNullSafeLessThan(fieldAx, fieldBy),
-		expression.NewLessThanOrEqual(fieldAx, fieldBy),
-		expression.NewNullSafeLessThanOrEqual(fieldAx, fieldBy),
-		expression.NewAnd(
-			expression.NewEquals(fieldAx, fieldBy),
-			expression.NewEquals(fieldAx, litOne),
-		),
-		expression.NewOr(
-			expression.NewEquals(fieldAx, fieldBy),
-			expression.NewEquals(litOne, litOne),
-		),
-		expression.NewNot(
-			expression.NewEquals(fieldAx, fieldBy),
-		),
-		expression.NewIsFalse(
-			expression.NewEquals(fieldAx, fieldBy),
-		),
-		expression.NewIsTrue(
-			expression.NewEquals(fieldAx, fieldBy),
-		),
-		expression.NewIsNull(
-			expression.NewEquals(fieldAx, fieldBy),
-		),
-	}
-
-	nonMatching := []sql.Expression{
-		expression.NewEquals(
-			expression.NewGetFieldWithTable(0, sql.Int64, "b", "x", false),
-			expression.NewGetFieldWithTable(0, sql.Int64, "b", "y", false),
-		),
-		expression.NewEquals(
-			expression.NewGetFieldWithTable(0, sql.Int64, "b", "x", false),
-			aggregation.NewMax(expression.NewGetFieldWithTable(0, sql.Int64, "b", "y", false)),
-		),
-	}
-
-	tests := make([]analyzerFnTestCase, 0, len(matching)+len(nonMatching))
-	for _, t := range matching {
-		new := analyzerFnTestCase{
-			name: t.String(),
-			node: plan.NewFilter(
-				t,
-				plan.NewCrossJoin(
-					plan.NewResolvedTable(tableA, nil, nil),
-					plan.NewTableAlias("b", plan.NewResolvedTable(tableB, nil, nil)),
-				),
-			),
-			expected: plan.NewInnerJoin(
-				plan.NewResolvedTable(tableA, nil, nil),
-				plan.NewTableAlias("b", plan.NewResolvedTable(tableB, nil, nil)),
-				t,
-			),
-		}
-		tests = append(tests, new)
-	}
-	for _, t := range nonMatching {
-		new := analyzerFnTestCase{
-			name: t.String(),
-			node: plan.NewFilter(
-				t,
-				plan.NewCrossJoin(
-					plan.NewResolvedTable(tableA, nil, nil),
-					plan.NewTableAlias("b", plan.NewResolvedTable(tableB, nil, nil)),
-				),
-			),
-			expected: plan.NewFilter(
-				t,
-				plan.NewCrossJoin(
-					plan.NewResolvedTable(tableA, nil, nil),
-					plan.NewTableAlias("b", plan.NewResolvedTable(tableB, nil, nil)),
-				),
-			),
-		}
-		tests = append(tests, new)
-	}
-
-	nested := []analyzerFnTestCase{
-		{
-			name: "nested cross joins",
-			node: plan.NewFilter(
-				expression.NewAnd(
-					expression.NewEquals(
-						fieldAx,
-						fieldBy,
-					),
-					expression.NewAnd(
-						expression.NewEquals(
-							expression.NewGetFieldWithTable(0, sql.Int64, "b", "x", false),
-							expression.NewGetFieldWithTable(1, sql.Int64, "c", "y", false),
-						),
-						expression.NewAnd(
-							expression.NewEquals(
-								expression.NewGetFieldWithTable(0, sql.Int64, "a", "x", false),
-								expression.NewGetFieldWithTable(0, sql.Int64, "a", "x", false),
-							),
-							expression.NewEquals(
-								expression.NewGetFieldWithTable(0, sql.Int64, "c", "x", false),
-								expression.NewGetFieldWithTable(1, sql.Int64, "d", "y", false),
-							),
-						),
-					),
-				),
-				plan.NewCrossJoin(
-					plan.NewResolvedTable(tableA, nil, nil),
-					plan.NewCrossJoin(
-						plan.NewTableAlias("b", plan.NewResolvedTable(tableB, nil, nil)),
-						plan.NewCrossJoin(
-							plan.NewTableAlias("c", plan.NewResolvedTable(tableB, nil, nil)),
-							plan.NewTableAlias("d", plan.NewResolvedTable(tableB, nil, nil)),
-						),
-						),
-				),
-			),
-			expected: plan.NewFilter(
-				expression.NewEquals(
-					expression.NewGetFieldWithTable(0, sql.Int64, "a", "x", false),
-					expression.NewGetFieldWithTable(0, sql.Int64, "a", "x", false),
-				),
-				plan.NewInnerJoin(
-					plan.NewResolvedTable(tableA, nil, nil),
-					plan.NewInnerJoin(
-						plan.NewTableAlias("b", plan.NewResolvedTable(tableB, nil, nil)),
-						plan.NewInnerJoin(
-							plan.NewTableAlias("c", plan.NewResolvedTable(tableB, nil, nil)),
-							plan.NewTableAlias("d", plan.NewResolvedTable(tableB, nil, nil)),
-							expression.NewEquals(
-								expression.NewGetFieldWithTable(0, sql.Int64, "c", "x", false),
-								expression.NewGetFieldWithTable(1, sql.Int64, "d", "y", false),
-							),
-						),
-						expression.NewEquals(
-							expression.NewGetFieldWithTable(0, sql.Int64, "b", "x", false),
-							expression.NewGetFieldWithTable(1, sql.Int64, "c", "y", false),
-						),
-						),
-					expression.NewEquals(fieldAx, fieldBy),
-				),
-			),
-		},
-	}
-	tests = append(tests, nested...)
-
-	runTestCases(t, sql.NewEmptyContext(), nested, NewDefault(sql.NewDatabaseProvider()), getRule("replace_cross_joins"))
 }
